teset b

+++ _binary.c
@@ -12,8 +12,8 @@
 int _b(const char *format, va_list arg, int *pcount)
 {
 	unsigned int num = va_arg(arg, unsigned int);
-	long unsigned int Tens = 1;
-	unsigned int result = 0;
+	long long unsigned int Tens = 1;
+	long long unsigned int result = 0;
 
 	(void)format;
 	while (num)
